requirements.txt
streamlit
pandas
bcrypt
openpyxl
psycopg2-binary